feat: Implement blame continuation from an existing checkpoint
In this commit, we introduce the ability to execute a blame operation from a previously generated checkpoint. This functionality makes the computation of incremental blame operations more efficient by reusing already computed information.

The refactoring performed in commit id d22965f8 allowed effortless integration of this feature into the existing blame processing algorithm.

The fundamental distinction between a regular blame and a blame operation from a checkpoint is that in the latter, we can pre-fill the `BlameState` with some `BlameEntry` instances, thereby reducing the number of `UnblamedHunk` instances we need to process.

This update’s algorithm for incorporating the detected modifications since the checkpoint into the blame entries is encapsulated in the `update_checkpoint_blames_with_changes` method.

The newly introduced `BlameCheckpoint` type is a public type that users can utilize to establish a checkpoint state for their blame operations.